Read env vars in task tests too.
