+++ test/query.js
@@ -3,10 +3,10 @@
 require('should');
 global.fetch = require('node-fetch');
 
-const home = require('os').homedir()
 const clientLib = require('../lib/client.js');
+const config = require('./config.js');
 
-const credentials = readCredentials()
+const credentials = config.readCredentials()
 const client = clientLib.create(credentials)
 const QUERY = 'from demo.ecommerce.data select eventdate,protocol,statusCode,method'
 const from = new Date(Date.now() - 60 * 1000)
@@ -59,22 +59,4 @@ describe('Browser client', () => {
     });
   })
 });
-
-function readCredentials() {
-  const env = {
-    apiKey: process.env.DEVO_KEY,
-    apiSecret: process.env.DEVO_SECRET,
-    token: process.env.DEVO_TOKEN,
-    url: process.env.DEVO_URL,
-  }
-  try {
-    const read = require(home + '/.devo.json')
-    Object.keys(env).forEach(key => {
-      if (env[key]) read[key] = env[key]
-    })
-    return read
-  } catch(exception) {
-    return env
-  }
-}
